Worked on the write path for multi-file projects.
